Fix Sample::from_svmlight.
I thought that the SVMLight format files are 1-indexed, and it is true for the files in LIBSVM site.
On the other hand, the sklearn.datasets.dump_svmlight_file function outputs, by default, 0-indexed SVMLight format file.
So, the previous version of this function panics since I designed to read 1-indexed file.
I fixed this issue.