Update the Booster trait to introduce graph-separation boosting.
The previous version of Booster trait only allows to return CombinedHypothesis<H>,
which is a majority vote of base hypotheses.

On the other hand, GraphSeparationBoosting, which is introduced in the next update,
adopts arbitrary aggregation that is consistent with training examples.

This update allows you to use other types of aggregation.